[ refact ] : 커뮤니티 좋아요 패키지를 커뮤니티 도메인 내부로 이동
