+++ src/main/java/com/talkit/app/domain/community/repository/CommentLikeRepository.java
@@ -1,6 +1,6 @@
-package com.talkit.app.domain.like.repository;
+package com.talkit.app.domain.community.repository;
 
-import com.talkit.app.domain.like.entity.CommentLike;
+import com.talkit.app.domain.community.entity.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
