+++ src/main/java/com/talkit/app/domain/community/service/CommunityService.java
@@ -8,8 +8,8 @@ import com.talkit.app.domain.community.dto.*;
 import com.talkit.app.domain.community.entity.Community;
 import com.talkit.app.domain.community.repository.CommunityCommentRepository;
 import com.talkit.app.domain.community.repository.CommunityRepository;
-import com.talkit.app.domain.like.entity.CommunityLike;
-import com.talkit.app.domain.like.repository.CommunityLikeRepository;
+import com.talkit.app.domain.community.entity.CommunityLike;
+import com.talkit.app.domain.community.repository.CommunityLikeRepository;
 import com.talkit.app.domain.user.entity.User;
 import com.talkit.app.domain.user.repository.UserRepository;
 import com.talkit.app.global.dto.PageRequestVO;
